requirements.txt
numpy
torch
soundfile
torchaudio
python-telegram-bot
